[tests] Fix failing tests
A few tests are failing when you install release-tools
as a package. This commit fixes all those failing tests.

This is necessary as we would soon upload automate the
release process using github actions and upload the package
to PyPI whenever there is a new release.

It also fixes a few flake8 errors.

Signed-off-by: Venu Vardhan Reddy Tekula <venu@bitergia.com>